+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- Changed an error log to a warning log.
+
 ## [v15.6.5](https://github.com/lexicalunit/spellbot/releases/tag/v15.6.5) - 2025-08-20
 
 ### Changed